+++ AutoMapRoom/obj/Debug/netstandard2.1/GeneratedModInfo.cs
@@ -6,6 +6,6 @@ namespace AutoMapRoom
     {
         public const string GUID = "s0apy.AutoMapRoom";
         public const string NAME = "AutoMapRoom";
-        public const string VERSION = "1.0.9";
+        public const string VERSION = "1.1.0";
     }
 }