+++ AutoMapRoom/_Thunderstore/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.9
 - True Field Name Priority
 - Ultra-Low Latency Debounce
+- Sex changelog
 
 ## 1.0.8
 - How many fixes can we do speedrun?