+++ AutoMapRoom/_Thunderstore/CHANGELOG.md
@@ -4,6 +4,7 @@
 - True Field Name Priority
 - Ultra-Low Latency Debounce
 - Sex changelog
+- pls cache pls
 
 ## 1.0.8
 - How many fixes can we do speedrun?