+++ AutoMapRoom/_Thunderstore/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Time-based Debounce
 - Sex changelog
 - pls cache pls
+- this is torture omg omlllll
 
 ## 1.0.8
 - How many fixes can we do speedrun?