+++ AutoMapRoom/_Thunderstore/CHANGELOG.md
@@ -8,6 +8,7 @@
 - this is torture omg omlllll
 - this is turning into badupdate xbox 360
 - culprit had to be debounce itself
+- holy spammer direct state check it btch danM!
 
 ## 1.0.8
 - How many fixes can we do speedrun?